feat: add working clock to second view #2
Took 9 minutes
